refactor: move indexer lookup into indexer namespace
