+++ src/index.ts
@@ -10,7 +10,6 @@ export {
 export type { Addressable, ReadableAddress } from '../packages/common/src/address'
 export * from './amount'
 export * from './config'
-export * as indexer from './indexer-lookup'
 export * from './transaction'
 export { AlgorandClient } from './types/algorand-client'
 export * from './types/debugging'
