+++ src/indexer-client/index.ts
@@ -1,1 +1,2 @@
 export * from '@algorandfoundation/algokit-indexer-client'
+export * from './indexer-lookup'
