+++ src/types/app-deployer.ts
@@ -3,7 +3,7 @@ import { Address, Expand, getAddress, getApplicationAddress, ReadableAddress } f
 import { IndexerClient } from '@algorandfoundation/algokit-indexer-client'
 import { TransactionType } from '@algorandfoundation/algokit-transact'
 import { Config } from '../config'
-import * as indexer from '../indexer-lookup'
+import { lookupAccountCreatedApplicationByAddress, searchTransactions } from '../indexer-client'
 import { calculateExtraProgramPages } from '../util'
 import { AlgorandClientTransactionSender } from './algorand-client-transaction-sender'
 import {
@@ -507,7 +507,7 @@ export class AppDeployer {
     }
 
     // Extract all apps that account created
-    const createdApps = (await indexer.lookupAccountCreatedApplicationByAddress(this._indexer, creatorString))
+    const createdApps = (await lookupAccountCreatedApplicationByAddress(this._indexer, creatorString))
       .map((a) => {
         return { id: a.id, createdAtRound: a.createdAtRound!, deleted: a.deleted }
       })
@@ -517,7 +517,7 @@ export class AppDeployer {
     const apps = await Promise.all(
       createdApps.map(async (createdApp) => {
         // Find any app transactions for that app in the round it was created (should always just be a single creation transaction)
-        const appTransactions = await indexer.searchTransactions(this._indexer!, {
+        const appTransactions = await searchTransactions(this._indexer!, {
           minRound: createdApp.createdAtRound,
           txType: TransactionType.AppCall,
           applicationId: createdApp.id,
